Guard OpenAI Agents scaffold tests behind an optional-dep importorskip
TestOpenAIAgentsToolCorrection (and its Agents-only helpers) need the optional
`agents`/`openai` deps, which are not installed in the default CI test groups,
so they broke the default test command. Move them into a dedicated
tests/core/harness/test_openai_agents_scaffold.py guarded at module level by
pytest.importorskip("agents") / ("openai"); test_scaffold.py now keeps only the
non-Agents TestGetScaffold and no longer imports agents/openai. This is a pure
test move plus import guard, with no test logic changed.

Addresses review feedback on #239.
